Allow clicking on settings FAB
