+++ lib/page/map/settings_controls_widget.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:insigno_frontend/page/settings/settings_page.dart';
 
 class SettingsControlsWidget extends StatelessWidget {
   final VoidCallback onFilterPressed;
@@ -21,7 +22,7 @@ class SettingsControlsWidget extends StatelessWidget {
             padding: const EdgeInsets.only(top: 8),
             child: FloatingActionButton(
               heroTag: "settings",
-              onPressed: () => {},
+              onPressed: () => Navigator.pushNamed(context, SettingsPage.routeName),
               tooltip: l10n.settings,
               mini: true,
               child: const Icon(Icons.settings),
